+++ src/components/dashboard/sidebar/SubsidiesSummary.jsx
@@ -8,9 +8,7 @@ import classNames from 'classnames';
 import CouponCodesSummaryCard from './CouponCodesSummaryCard';
 import SubscriptionSummaryCard from './SubscriptionSummaryCard';
 import LearnerCreditSummaryCard from './LearnerCreditSummaryCard';
-import { LICENSE_STATUS } from '../../enterprise-user-subsidy/data/constants';
 import SidebarCard from './SidebarCard';
-import { getOfferExpiringFirst, getPolicyExpiringFirst } from './utils';
 import {
   useBrowseAndRequest,
   useCouponCodes,
@@ -18,25 +16,10 @@ import {
   useEnterpriseCustomer,
   useEnterpriseOffers,
   useIsAssignmentsOnlyLearner,
-  useRedeemablePolicies,
   useSubscriptions,
 } from '../../app/data';
 import { COURSE_STATUSES } from '../../../constants';
-
-function getLearnerCreditSummaryCardData({ enterpriseOffers, redeemableLearnerCreditPolicies }) {
-  const learnerCreditPolicyExpiringFirst = getPolicyExpiringFirst(redeemableLearnerCreditPolicies?.redeemablePolicies);
-  const enterpriseOfferExpiringFirst = getOfferExpiringFirst(enterpriseOffers);
-
-  if (!learnerCreditPolicyExpiringFirst && !enterpriseOfferExpiringFirst) {
-    return undefined;
-  }
-
-  return {
-    expirationDate: (
-      learnerCreditPolicyExpiringFirst?.subsidyExpirationDate || enterpriseOfferExpiringFirst?.endDatetime
-    ),
-  };
-}
+import { useHasActiveSubsidy } from '../../hooks';
 
 const SubsidiesSummary = ({
   className,
@@ -51,17 +34,16 @@ const SubsidiesSummary = ({
   const { data: subscriptions } = useSubscriptions();
   const { data: couponCodes } = useCouponCodes();
   const { data: enterpriseOffersData } = useEnterpriseOffers();
-  const { data: redeemableLearnerCreditPolicies } = useRedeemablePolicies();
   const { data: { requests } } = useBrowseAndRequest();
+  const {
+    hasAvailableSubsidyOrRequests,
+    hasAvailableLearnerCreditPolicies,
+    hasAssignedCodesOrCodeRequests,
+    hasActiveLicenseOrLicenseRequest,
+    learnerCreditSummaryCardData,
+  } = useHasActiveSubsidy();
 
   const isAssignmentOnlyLearner = useIsAssignmentsOnlyLearner();
-
-  const learnerCreditSummaryCardData = getLearnerCreditSummaryCardData({
-    enterpriseOffers: enterpriseOffersData.enterpriseOffers,
-    redeemableLearnerCreditPolicies,
-  });
-
-  // const { requestsBySubsidyType } = useContext(SubsidyRequestsContext);
 
   // if there are course enrollments, the cta button below will be the only one on the page
   const ctaButtonVariant = useMemo(() => {
@@ -75,21 +57,6 @@ const SubsidiesSummary = ({
       .flat().length > 0;
     return hasCourseEnrollments ? 'primary' : 'outline-primary';
   }, [allEnrollmentsByStatus]);
-
-  const hasActiveLicenseOrLicenseRequest = (
-    subscriptions.subscriptionLicense?.status === LICENSE_STATUS.ACTIVATED
-    || requests.subscriptionLicenses.length > 0
-  );
-
-  const hasAssignedCodesOrCodeRequests = (
-    couponCodes.couponCodeAssignments.length > 0
-    || requests.couponCodes.length > 0
-  );
-  const hasAvailableLearnerCreditPolicies = redeemableLearnerCreditPolicies?.redeemablePolicies.length > 0;
-
-  const hasAvailableSubsidyOrRequests = (
-    hasActiveLicenseOrLicenseRequest || hasAssignedCodesOrCodeRequests || learnerCreditSummaryCardData
-  );
 
   if (!hasAvailableSubsidyOrRequests) {
     return null;
