+++ src/components/hooks/index.js
@@ -2,3 +2,4 @@ export { default as useEntepriseFeatures } from './useEnterpriseFeatures';
 export { default as useCatalogsForSubsidyRequests } from './useCatalogsForSubsidyRequests';
 export { default as useContentHighlights } from './useContentHighlights';
 export { default as useAcademies } from './useAcademies';
+export { default as useHasActiveSubsidy } from './useHasActiveSubsidy';
